+++ src/euphorie/upgrade/deployment/v1/20211216172426_toggle_for_self_registration/upgrade.py
@@ -1,0 +1,8 @@
+from ftw.upgrade import UpgradeStep
+
+
+class ToggleForSelfRegistration(UpgradeStep):
+    """Toggle for self registration."""
+
+    def __call__(self):
+        self.install_upgrade_profile()